OpenCV align left right to blue tower goal
